--- src/App.tsx
@@ -1,10 +1,0 @@
-import './styles/base.css'
-import 'lume'
-import Main from './views/main/Main'
-
-
-function App() {
-  return <Main/>
-}
-
-export default App
