--- src/config.ts
@@ -1,13 +1,0 @@
-/** * Main scene dimensions in pixels.  */
-export const SCENE_DIMENSIONS = {
-    width: 940,
-    height: 600
-}
-
-export const INITIAL_SCENE = "Elevator";
-
-export const FOV = 45;
-
-export const DEFAULT_DIALOGUE_SENDER = "Arda";
-
-export const LOGIN_MESSAGE = "INSTANCE INIT | [ DAEMON-GARDEN v.0.400 ]"
