+++ src/config/init.ts
@@ -1,0 +1,3 @@
+export const INITIAL_SCENE = "Elevator";
+export const LOGIN_MESSAGE = "INSTANCE INIT | [ DAEMON-GARDEN v.0.400 ]"
+export const DEFAULT_DIALOGUE_SENDER = "Arda";
