+++ src/config/lume.ts
@@ -1,0 +1,2 @@
+
+export const FOV = 45;
