+++ src/config/ui.ts
@@ -1,0 +1,5 @@
+/** * Main scene dimensions in pixels.  */
+export const SCENE_DIMENSIONS = {
+    width: 940,
+    height: 600
+};
