+++ src/index.tsx
@@ -1,7 +1,9 @@
 /* @refresh reload */
 import { render } from 'solid-js/web'
-import App from './App.tsx'
+import 'lume'
+import Main from './views/main/Main.tsx'
+import '@/styles/base.css'
 
 const root = document.getElementById('root')
 
-render(() => <App />, root!)
+render(() => <Main/>, root!)
