+++ src/tests/dialogue.test.ts
@@ -9,7 +9,7 @@ import {
   VISUALIZER,
 } from "@/core/dialogue/dialogueNode";
 import { DialogueContext, DialogueNode } from "@/core/dialogue/dialogueNode.types";
-import { DEFAULT_DIALOGUE_SENDER } from "@/config";
+import { DEFAULT_DIALOGUE_SENDER } from "@/config/init";
 
 describe("DialogueNode helpers", () => {
   it("createDialogueNode basics and id uniqueness", () => {
