+++ src/views/main/ui/EventLog.tsx
@@ -3,7 +3,7 @@ import br from "@/assets/ui/corners/da/br.png"
 import tr from "@/assets/ui/corners/da/tr.png"
 import tl_el from "../assets/tl_el.png"
 import { createEffect, on, createSignal, For } from "solid-js";
-import { LOGIN_MESSAGE } from "@/config";
+import { LOGIN_MESSAGE } from "@/config/init";
 
 const [logMessages, setLogMessages] = createSignal<{ id: number, text: string, color: string }[]>([
     {id: 0, text: LOGIN_MESSAGE, color: '#cfb886ff'}
