+++ src/views/main/ui/SceneContainer.tsx
@@ -3,7 +3,8 @@ import bl_scene from "../assets/bl_scene.png"
 import br from "@/assets/ui/corners/da/br.png"
 import tl from "@/assets/ui/corners/da/tl.png"
 import tr from "@/assets/ui/corners/da/tr.png"
-import { INITIAL_SCENE, SCENE_DIMENSIONS } from "@/config";
+import { INITIAL_SCENE } from "@/config/init";
+import { SCENE_DIMENSIONS } from "@/config/ui";
 import { createEffect, createSignal, ErrorBoundary, on, Show, Suspense } from "solid-js";
 import { Dynamic } from "solid-js/web";
 import { DialogueService } from "@/core/dialogue/dialogueService";
