+++ src/views/main/ui/SceneMenu/SceneMenuWrapper.tsx
@@ -1,5 +1,5 @@
 import { ParentProps, createSignal } from "solid-js";
-import { SCENE_DIMENSIONS } from "@/config";
+import { SCENE_DIMENSIONS } from "@/config/ui";
 import { MenuOption, SceneContextMenu } from "./scenemenu.types";
 import { SceneMenuContext } from "./SceneMenuContext";
 import SceneMenu from "./SceneMenu";
